+++ src/components/analysis/parsing/lr1_automaton_component.jsx
@@ -7,7 +7,7 @@ export default function({ grammar }) {
   return (
     <section id={ID} className="analysis">
       <h2>{TITLE}</h2>
-      <AbstractLRAutomatonComponent grammar={grammar} automaton={grammar.calculations.lr1Automaton} title={TITLE} />
+      <AbstractLRAutomatonComponent grammar={grammar} automaton={grammar.calculations.lr1Automaton} />
     </section>
   );
 }